test(nextcloud): split webdav tests
