+++ packages/nextcloud/lib/src/api/webdav/models/props.dart
@@ -2,7 +2,6 @@
 // coverage:ignore-file
 import 'package:meta/meta.dart';
 import 'package:nextcloud/src/api/webdav/webdav.dart';
-import 'package:nextcloud/src/utils/date_time.dart';
 import 'package:timezone/timezone.dart' as tz;
 import 'package:xml/xml.dart';
 import 'package:xml_annotation/xml_annotation.dart' as annotation;
