+++ packages/nextcloud/lib/src/api/webdav/utils/utils.dart
@@ -1,4 +1,5 @@
 export 'options_parser.dart';
 export 'webdav_response_converter.dart';
 export 'webdav_uri.dart';
+export 'xml_date_time_converter.dart';
 export 'xml_duration_converter.dart';
