+++ packages/nextcloud/lib/webdav.dart
@@ -2,7 +2,14 @@ import 'package:nextcloud/src/api/webdav/webdav_client.dart';
 import 'package:nextcloud/src/nextcloud_client.dart';
 
 export 'package:nextcloud/src/api/webdav/webdav.dart'
-    hide DurationXMLConverter, WebDavCSRFClient, constructUri, parseWebDavOptions;
+    hide
+        DurationXMLConverter,
+        HttpDateXMLConverter,
+        ISO8601XMLConverter,
+        UnixEpochXMLConverter,
+        WebDavCSRFClient,
+        constructUri,
+        parseWebDavOptions;
 
 /// Client for WebDAV.
 extension WebDAVExtension on NextcloudClient {
