+++ packages/nextcloud/test/api/webdav/webdav_test.dart
@@ -4,7 +4,6 @@ import 'dart:typed_data';
 
 import 'package:mocktail/mocktail.dart';
 import 'package:nextcloud/nextcloud.dart';
-import 'package:nextcloud/src/api/webdav/webdav.dart';
 import 'package:nextcloud/src/utils/date_time.dart';
 import 'package:nextcloud/webdav.dart';
 import 'package:nextcloud_test/nextcloud_test.dart';
@@ -16,132 +15,6 @@ class MockCallbackFunction extends Mock {
 }
 
 void main() {
-  group('constructUri', () {
-    for (final values in [
-      ('http://cloud.example.com', 'http://cloud.example.com'),
-      ('http://cloud.example.com/', 'http://cloud.example.com'),
-      ('http://cloud.example.com/subdir', 'http://cloud.example.com/subdir'),
-      ('http://cloud.example.com/subdir/', 'http://cloud.example.com/subdir'),
-    ]) {
-      final baseURL = Uri.parse(values.$1);
-      final sanitizedBaseURL = Uri.parse(values.$2);
-
-      test(baseURL, () {
-        expect(
-          constructUri(baseURL).toString(),
-          '$sanitizedBaseURL$webdavBase',
-        );
-        expect(
-          constructUri(baseURL, PathUri.parse('/')).toString(),
-          '$sanitizedBaseURL$webdavBase',
-        );
-        expect(
-          constructUri(baseURL, PathUri.parse('test')).toString(),
-          '$sanitizedBaseURL$webdavBase/test',
-        );
-        expect(
-          constructUri(baseURL, PathUri.parse('test/')).toString(),
-          '$sanitizedBaseURL$webdavBase/test',
-        );
-        expect(
-          constructUri(baseURL, PathUri.parse('/test')).toString(),
-          '$sanitizedBaseURL$webdavBase/test',
-        );
-        expect(
-          constructUri(baseURL, PathUri.parse('/test/')).toString(),
-          '$sanitizedBaseURL$webdavBase/test',
-        );
-      });
-    }
-  });
-
-  group('PathUri', () {
-    test('isAbsolute', () {
-      expect(PathUri.parse('').isAbsolute, false);
-      expect(PathUri.parse('/').isAbsolute, true);
-      expect(PathUri.parse('test').isAbsolute, false);
-      expect(PathUri.parse('test/').isAbsolute, false);
-      expect(PathUri.parse('/test').isAbsolute, true);
-      expect(PathUri.parse('/test/').isAbsolute, true);
-    });
-
-    test('isDirectory', () {
-      expect(PathUri.parse('').isDirectory, true);
-      expect(PathUri.parse('/').isDirectory, true);
-      expect(PathUri.parse('test').isDirectory, false);
-      expect(PathUri.parse('test/').isDirectory, true);
-      expect(PathUri.parse('/test').isDirectory, false);
-      expect(PathUri.parse('/test/').isDirectory, true);
-    });
-
-    test('pathSegments', () {
-      expect(PathUri.parse('').pathSegments, isEmpty);
-      expect(PathUri.parse('/').pathSegments, isEmpty);
-      expect(PathUri.parse('test').pathSegments, ['test']);
-      expect(PathUri.parse('test/').pathSegments, ['test']);
-      expect(PathUri.parse('/test').pathSegments, ['test']);
-      expect(PathUri.parse('/test/').pathSegments, ['test']);
-    });
-
-    test('path', () {
-      expect(PathUri.parse('').path, '');
-      expect(PathUri.parse('/').path, '/');
-      expect(PathUri.parse('test').path, 'test');
-      expect(PathUri.parse('test/').path, 'test/');
-      expect(PathUri.parse('/test').path, '/test');
-      expect(PathUri.parse('/test/').path, '/test/');
-    });
-
-    test('normalization', () {
-      expect(PathUri.parse('/test/abc/').path, '/test/abc/');
-      expect(PathUri.parse('//test//abc//').path, '/test/abc/');
-      expect(PathUri.parse('///test///abc///').path, '/test/abc/');
-    });
-
-    test('name', () {
-      expect(PathUri.parse('').name, '');
-      expect(PathUri.parse('test').name, 'test');
-      expect(PathUri.parse('/test/').name, 'test');
-      expect(PathUri.parse('abc/test').name, 'test');
-      expect(PathUri.parse('/abc/test/').name, 'test');
-    });
-
-    test('parent', () {
-      expect(PathUri.parse('').parent, null);
-      expect(PathUri.parse('/').parent, null);
-      expect(PathUri.parse('test').parent, PathUri.parse(''));
-      expect(PathUri.parse('test/abc').parent, PathUri.parse('test/'));
-      expect(PathUri.parse('test/abc/').parent, PathUri.parse('test/'));
-      expect(PathUri.parse('/test/abc').parent, PathUri.parse('/test/'));
-      expect(PathUri.parse('/test/abc/').parent, PathUri.parse('/test/'));
-    });
-
-    test('join', () {
-      expect(PathUri.parse('').join(PathUri.parse('test')), PathUri.parse('test'));
-      expect(PathUri.parse('/').join(PathUri.parse('test')), PathUri.parse('/test'));
-      expect(() => PathUri.parse('test').join(PathUri.parse('abc')), throwsA(isA<StateError>()));
-      expect(PathUri.parse('test/').join(PathUri.parse('abc')), PathUri.parse('test/abc'));
-      expect(PathUri.parse('test/').join(PathUri.parse('abc/123')), PathUri.parse('test/abc/123'));
-      expect(PathUri.parse('/test/').join(PathUri.parse('abc')), PathUri.parse('/test/abc'));
-      expect(PathUri.parse('/test/').join(PathUri.parse('/abc')), PathUri.parse('/test/abc'));
-      expect(PathUri.parse('/test/').join(PathUri.parse('/abc/')), PathUri.parse('/test/abc/'));
-    });
-
-    test('rename', () {
-      expect(PathUri.parse('').rename('test'), PathUri.parse(''));
-      expect(PathUri.parse('test').rename('abc'), PathUri.parse('abc'));
-      expect(PathUri.parse('test/').rename('abc'), PathUri.parse('abc/'));
-      expect(PathUri.parse('test/abc').rename('123'), PathUri.parse('test/123'));
-      expect(PathUri.parse('test/abc/').rename('123'), PathUri.parse('test/123/'));
-      expect(() => PathUri.parse('test').rename('abc/'), throwsA(isA<Exception>()));
-      expect(() => PathUri.parse('test/').rename('abc/'), throwsA(isA<Exception>()));
-      expect(() => PathUri.parse('test').rename('/abc'), throwsA(isA<Exception>()));
-      expect(() => PathUri.parse('test/').rename('/abc'), throwsA(isA<Exception>()));
-      expect(() => PathUri.parse('test').rename('abc/123'), throwsA(isA<Exception>()));
-      expect(() => PathUri.parse('test/').rename('abc/123'), throwsA(isA<Exception>()));
-    });
-  });
-
   test('Chunked responses', () async {
     await HttpServer.bind('127.0.0.1', 0).then((server) async {
       server.listen((request) {
@@ -186,7 +59,7 @@ void main() {
           )
           .forEach(buffer.add);
 
-      expect(buffer.toBytes(), Uint8List.fromList(utf8.encode('123')));
+      expect(buffer.toBytes(), utf8.encode('123'));
       expect(progress, [1]);
     });
   });
